[fix]サービスの説明部分をコンポーネント化
